+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## next
+
+- Experimental ES6/Harmony mode based on UglifyJS2 Harmony branch
+
 ## 3.1.13 (3 April 2017)
 
 - update UglifyJS to 2.8.21